Code/camera/requirements.txt
opencv-python==4.5.3.56
paho-mqtt==1.6.1
requests==2.26.0